+++ app/Http/Controllers/AdminController.php
@@ -4,19 +4,14 @@ namespace App\Http\Controllers;
 
 use App\Admin;
 use App\Http\Requests\AdminRequest;
+use Illuminate\Database\QueryException;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\MessageBag;
+use Mockery\Exception;
 
 
 class AdminController extends Controller
 {
-    public function __construct()
-    {
-        if(Auth::user() == null)
-            Auth::shouldUse("web_admin");
-
-        $this->middleware('auth');
-    }
-
     /**
      * Display a listing of the resource.
      *
@@ -25,16 +20,25 @@ class AdminController extends Controller
 
     public function admin()
     {
+        if(!$this->testLogin())
+            return redirect()->route("login");
+
         return view('admin')->with("admins",Admin::all());
     }
 
     public function getFormAdmin($id=null)
     {
-            return isset($id) ? view('admin.new')->with("admin", Admin::find($id)) : view('admin.new')->with("admin", new Admin());
+        if(!$this->testLogin())
+            return redirect()->route("login");
+
+        return isset($id) ? view('admin.new')->with("admin", Admin::find($id)) : view('admin.new')->with("admin", new Admin());
     }
 
     public function postFormAdmin(AdminRequest $request,$id=null)
     {
+        if(!$this->testLogin())
+            return redirect()->route("login");
+
         if(isset($id))
             $admin = Admin::find($id);
         else
@@ -52,15 +56,31 @@ class AdminController extends Controller
 
     public function activeAdmin($id)
     {
+        if(!$this->testLogin())
+            return redirect()->route("login");
+
         $admin=Admin::find($id);
         $admin->active =! $admin->active;
         $admin->save();
-        return redirect()->route('hotesse');
+        return redirect()->route('admin');
     }
 
     public function deleteAdmin($id)
     {
-        Admin::find($id)->delete();
-        return redirect()->route('hotesse');
+        if(!$this->testLogin())
+            return redirect()->route("login");
+
+        try {
+            Admin::find($id)->delete();
+            return redirect()->route('admin');
+        }
+        catch (QueryException $e)
+        {
+            $bag = new MessageBag();
+            $bag->add("err","une erreur s'est produite pendant la suppression");
+
+            return redirect()->route("getUpdateAdmin",["id"=>$id])->withErrors($bag);
+        }
+
     }
 }
